backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

flask==2.3.3
flask-cors==5.0.1
moviepy==1.0.3
pydub==0.25.1
pytubefix==8.12.1
tqdm==4.65.0
google.genai
python-dotenv
pywebview[cef]
pyinstaller
